requirements.txt
pyyaml
discord.py